requirements.txt
dj-database-url==0.5.0
Django==2.2
django-ipware==2.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
psycopg2-binary==2.7.7
python-slugify==3.0.2
pytz==2018.9
sqlparse==0.3.0
text-unidecode==1.2
whitenoise==4.1.2
